Fix edge cases around volatile sources
In attempting to fix the integration tests I identified two bugs.

The first was that while evaluating formulas with volatile sources, the
first read would not promote to non-volatile since it wasn't yet watched
(still gathering dependencies). I fixed that by adding
a `isInWatcherContext` flag.

The second was that sources were being promoted to non-volatile signals
even if they didn't provide a subscriber. Obviously if we can't
subscribe to the value, we don't know when it changes so the source must
be volatile.

There's still one failing test. I'll get to that one next.